feature: add use case SearchAvailableItemsService

+++ src/main/java/com/lps/vitalMagic/inventory/application/view/ItemAttributeView.java
@@ -3,5 +3,5 @@ package com.lps.vitalMagic.inventory.application.view;
 public record ItemAttributeView(
         Long id,
         String name,
-        Byte value
+        int value
 ) {}
